init commit
Created a background with texts as the page introduction.
Connected to the API and, depending on the selected inputs, I appropriately fetch data and display it in a table.
Circumvented the limit of 100 tags per request and increased the number of tags per page to 300 instead of 100.
Handled errors and added a loading animation for components in the table.
Decided to use MobX for state management due to its simplicity, which is an advantage in less logically complex projects.#